chat-chien-detection/requirements.txt
tensorflow==2.20.0
numpy==1.26.0
pillow==10.0.0
matplotlib==3.7.1
scikit-learn==1.3.0
kaggle==1.5.16